-Aggiunti tutti i metodi necessari ad Evento.java -Aggiunti tutti i metodi per la gestione eventi necessari ad ControlloreUtente -Aggiunti tutti i metodi necessari per la gestione eventi ad ControlloreLocale -Aggiunti tab eventi e news nel dettaglio locale -Creato EventBean.java -Aggiunto form per aggiungere nuovo evento(tasto aggiungi non funzionante...) TODO: -stesso lavoro per le news -query di ricerca per gruppi di locali -funzione che cancelli i vecchi eventi

+++ branchJSF/luncharound/luncharound-ejb/src/java/luncharoundpkg/EventoFacade.java
@@ -44,7 +44,7 @@ public class EventoFacade extends AbstractFacade<Evento> implements EventoFacade
     }
     
     @Override
-    public List<Evento> findByLocali(List<Long> idLocali ){
+    public List<Evento> findByLocali(List<Integer> idLocali ){
         String selectQuery = "SELECT V FROM Evento V ORDER BY V.dataInizio, V.dataFine WHERE V.idLocale IN ?1";
 	Query searchById = em.createQuery(selectQuery);
 	searchById.setParameter(1, idLocali);
